+++ administrator/components/com_crowdfunding/CHANGELOG.md
@@ -10,8 +10,8 @@ Crowdfunding Platform Changelog
 * Fixed an issue with add a country.
 * Improved the performance. Now, the system makes fewer database calls and works faster.
 * Fixed and improved all modules, plug-ins and components that works with the crowdfunding platform.
-* It was improved the transaction process. The payment plugins work faster and use less memory.
-* It was added TransactionManager that provides to events handled by observable objects - onBeforeProcessTransaction and onAfterProcessTransaction.
+* It was improved the transaction process. The payment plugins work faster, use less memory and makes fewer database calls.
+* It was added TransactionManager that provides events handled by observable objects - onBeforeProcessTransaction and onAfterProcessTransaction.
 * Payment plugins use observable object to process transaction. 
 
 ###v2.4.1